chore: ViewModel 만들어만 두고 ListFragment의 Spinner에 motion layout 적용 초안 일단 완성
