+++ app/src/main/res/drawable/spinner_corner_border_box.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android">
+    <corners android:radius="12dp" />
+    <stroke
+        android:width="1dp"
+        android:color="@color/black" />
+    <solid android:color="@color/white" />
+</shape>
