+++ app/src/main/res/xml/motion_scene.xml
@@ -1,0 +1,43 @@
+<?xml version="1.0" encoding="utf-8"?>
+<MotionScene xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+
+    <Transition
+        app:constraintSetEnd="@+id/secondAnimation"
+        app:constraintSetStart="@+id/firstAnimation"
+        app:duration="500">
+        <OnClick
+            app:clickAction="transitionToEnd"
+            app:targetId="@+id/tvSpinner" />
+    </Transition>
+
+    <Transition
+        app:constraintSetEnd="@id/firstAnimation"
+        app:constraintSetStart="@id/secondAnimation"
+        app:duration="500">
+        <OnClick
+            app:clickAction="transitionToEnd"
+            app:targetId="@+id/tvSpinner"/>
+    </Transition>
+
+    <ConstraintSet android:id="@+id/firstAnimation">
+        <Constraint
+            android:id="@+id/memoList_cl_spinner"
+            android:layout_width="wrap_content"
+            android:layout_height="48dp"
+            android:layout_marginStart="20dp"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toTopOf="parent" />
+    </ConstraintSet>
+
+    <ConstraintSet android:id="@+id/secondAnimation">
+        <Constraint
+            android:id="@+id/memoList_cl_spinner"
+            android:layout_width="wrap_content"
+            android:layout_height="144dp"
+            android:layout_marginStart="20dp"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toTopOf="parent" />
+    </ConstraintSet>
+
+</MotionScene>
